+++ NEWS.md
@@ -1,3 +1,5 @@
+# microinverterdata 0.1.3
+
 # microinverterdata 0.1.2
 
 * Add mocked http data for APSystems and Fronius #2